backendserver/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
curl_cffi==0.12.0
Deprecated==1.2.18
fastapi==0.116.1
Flask==3.1.1
flask-cors==6.0.1
frozendict==2.4.6
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.4.0
MarkupSafe==3.0.2
multitasking==0.0.11
numpy==2.3.1
packaging==25.0
pandas==2.3.1
peewee==3.18.2
platformdirs==4.3.8
protobuf==6.31.1
psycopg==3.2.9
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yfinance==0.2.65
